requirements.txt
annotated-types==0.7.0
anyio==4.4.0
build==1.2.1
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
noise==1.2.2
nPerlinNoise==0.1.4a0
numexpr==2.10.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
perlin-noise==1.13
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rdflib==7.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sparql-dataframe==0.4
SPARQLWrapper==2.0.0
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
